requirements.txt
discord.py>=2.4.0,<3.0.0
requests>=2.32.0,<3.0.0
ecdsa>=0.19.0,<0.20.0
base58>=2.1.1,<3.0.0
web3>=7.0.0,<8.0.0
cryptography>=42.0.0,<46.0.0
python-dotenv>=1.0.1,<2.0.0
